🕹️ Control + Controllers: Update documentation
- Address all warnings when building documentation
- Add docc comments to all public APIs in both libraries
    - Fix docc symbol links/references
    - Cannot reference symbols across libraries: https://github.com/swiftlang/swift-docc/issues/208
- Move MPMusicPlaybackState extension to separate file
- Move SPTAppRemote extension to separate file
