+++ Sources/Control/Extensions/MPMusicPlaybackState.swift
@@ -1,0 +1,13 @@
+//
+//  MPMusicPlaybackState.swift
+//  ControlKit
+//
+
+import MediaPlayer
+
+extension MPMusicPlaybackState {
+    
+    var isPlaying: Bool {
+        [.playing, .seekingForward, .seekingBackward].contains(self)
+    }
+}
