+++ Sources/Control/Haptics.swift
@@ -7,9 +7,12 @@ import UIKit
 
 public extension Control {
     
+    /// 🫨 Control device haptics.
     @MainActor
     enum Haptics {
         
+        /// This function uses `UIImpactFeedbackGenerator/impactOccurred` to create a soft vibration,
+        /// which can be used to provide tactile feedback in response to user interactions.
         public static func vibrate() {
             UIImpactFeedbackGenerator().impactOccurred()
         }
